Made changes to enemies slope control as well
There is a bug where many short surfaces on their side (mainly the one in test area) will cause some jumps in rotation, mainly from changing modes of rotation a lot (3 times). Solutions can be check to make sure that raycasts are below the player (ones above shouldnt affect rotation). This doesnt fix the initial rotation jump (from leveling out on top of the slope to being the same slope it was. This is caused by it reading the underside of the slope and changing the scan range)